+++ ante/fee.go
@@ -50,7 +50,7 @@ func checkFees(minGasPrices sdk.DecCoins, fees sdk.Coins, gasLimit uint64) error
 		}
 	} else {
 		// in deliverTx = use protocol min gas price
-		minGasPrices = sdk.DecCoins{appparams.MinMinGasPrice}
+		minGasPrices = sdk.DecCoins{appparams.ProtocolMinGasPrice}
 	}
 
 	requiredFees := make(sdk.Coins, len(minGasPrices))
@@ -106,15 +106,15 @@ func IsOracleOrGravityTx(msgs []sdk.Msg) bool {
 // the required by protocol.
 func AssertMinProtocolGasPrice(gasPrices sdk.DecCoins) error {
 	for _, c := range gasPrices {
-		if c.Denom == appparams.MinMinGasPrice.Denom {
-			if c.Amount.LT(appparams.MinMinGasPrice.Amount) {
+		if c.Denom == appparams.ProtocolMinGasPrice.Denom {
+			if c.Amount.LT(appparams.ProtocolMinGasPrice.Amount) {
 				break // go to error below
 			}
 			return nil
 		}
 	}
 	return sdkerrors.ErrInsufficientFee.Wrapf(
-		"gas price too small; got: %v required min: %v", gasPrices, appparams.MinMinGasPrice)
+		"gas price too small; got: %v required min: %v", gasPrices, appparams.ProtocolMinGasPrice)
 }
 
 // getTxPriority returns naive tx priority based on the lowest fee amount (regardless of the